Extract 'extras' array helpers to a trait.
