+++ app/Feeds/Feed.php
@@ -2,7 +2,7 @@
 
 namespace App\Feeds;
 
-use App\Utilities\Arr;
+use App\Feeds\HasExtrasArray;
 use App\Utilities\Xml;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Collection;
@@ -10,6 +10,8 @@ use SimpleXMLElement;
 
 abstract class Feed
 {
+    use HasExtrasArray;
+
     /**
      * @var Collection
      */
@@ -63,50 +65,6 @@ abstract class Feed
     public function getEntries(): Collection
     {
         return empty($this->entries) ? collect() : $this->entries;
-    }
-
-    /**
-     * @var array
-     */
-    protected $extra = [];
-
-    /**
-     * Retrieve extra information that may have been parsed from the XML.
-     *
-     * @param string|null $key
-     * @param mixed $default
-     * @return mixed
-     */
-    public function getExtra($key = null, $default = null): mixed
-    {
-        if ($key) {
-            return data_get($this->extra, $key, $default);
-        }
-
-        return $this->extra;
-    }
-
-    /**
-     * Check to see if a key is available in the extras array.
-     *
-     * @param string $key
-     * @return boolean
-     */
-    public function hasExtra($key)
-    {
-        return Arr::has($this->extra, $key);
-    }
-
-    /**
-     * Record an 'extra' value that was parsed from the XML.
-     *
-     * @param string $key
-     * @param mixed $value
-     * @return void
-     */
-    public function setExtra($key, $value): void
-    {
-        $this->extra = data_set($this->extra, $key, $value);
     }
 
     /**
